# refine the performance of gather Op