<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->
Performance optimization
### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->
OPs
### Describe
<!-- Describe what this PR does -->
This PR
paddle.gather
```

run command: nvprof /workspace/tmp/anaconda2/envs/py37/bin/python /workspace/work/benchmark/api/tests_v2/gather.py --task speed --framework paddle --json_file /workspace/work/benchmark/api/tests_v2/configs/gather.json --config_id 0 --check_output False --profiler none --backward False --use_gpu True --repeat 5000 --allow_adaptive_repeat False --log_level 0
            Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:   99.71%  7.8124ms      5001  1.5620us  1.5350us  7.2320us  void paddle::operators::GatherGPUKernel<float, int>(float const *, int const *, paddle::operators::GatherGPUKernel<float, int>*, int, int, int, int, int)
                    0.21%  16.160us         9  1.7950us  1.7280us  2.0480us  [CUDA memcpy HtoD]

percent: 1.00; gpu_time: 7.8124 ms; calls: 5001; function: void paddle::operators::GatherGPUKernel<float, int>(float const *, int const *, paddle::operators::GatherGPUKernel<float, int>*, int, int, int, int, int)
total gpu_time: 7.8351 ms

```
develop
paddle.gather

```
run command: nvprof /workspace/tmp/anaconda2/envs/py37/bin/python /workspace/work/benchmark/api/tests_v2/gather.py --task speed --framework paddle --json_file /workspace/work/benchmark/api/tests_v2/configs/gather.json --config_id 0 --check_output False --profiler none --backward False --use_gpu True --repeat 5000 --allow_adaptive_repeat False --log_level 0
            Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:   39.96%  9.4163ms      5001  1.8820us  1.8240us  10.656us  [CUDA memcpy DtoH]
                   33.01%  7.7776ms      5001  1.5550us  1.5350us  7.2640us  void paddle::operators::GatherGPUKernel<float, int>(float const *, int const *, paddle::operators::GatherGPUKernel<float, int>*, int, int, int, int, int)
                   26.93%  6.3463ms      5001  1.2680us  1.1840us  7.1040us  void Eigen::internal::EigenMetaKernel<Eigen::TensorEvaluator<Eigen::TensorAssignOp<Eigen::TensorMap<Eigen::Tensor<long, int=1, int=1, long>, int=0, Eigen::MakePointer>, Eigen::TensorCwiseNullaryOp<Eigen::internal::scalar_constant_op<long>, Eigen::TensorMap<Eigen::Tensor<long, int=1, int=1, long>, int=0, Eigen::MakePointer> const > const > const , Eigen::GpuDevice>, long>(long, int=1)
                    0.07%  16.128us         9  1.7920us  1.6960us  2.0800us  [CUDA memcpy HtoD]

```

paddle.fluid.layers.gather
```
run command: nvprof /workspace/tmp/anaconda2/envs/py37/bin/python /workspace/work/benchmark/api/tests/gather.py --task speed --framework paddle --json_file /workspace/work/benchmark/api/tests/configs/gather.json --config_id 0 --check_output False --profiler none --backward False --use_gpu True --repeat 5000 --allow_adaptive_repeat False --log_level 0
            Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:   99.69%  7.5451ms      5001  1.5080us  1.4710us  7.3590us  void paddle::operators::GatherCUDAKernel<float, int>(float const *, int const *, paddle::operators::GatherCUDAKernel<float, int>*, unsigned long, unsigned long)
                    0.22%  16.896us         9  1.8770us  1.6960us  2.4000us  [CUDA memcpy HtoD]

percent: 1.00; gpu_time: 7.5451 ms; calls: 5001; function: void paddle::operators::GatherCUDAKernel<float, int>(float const *, int const *, paddle::operators::GatherCUDAKernel<float, int>*, unsigned long, unsigned long)
total gpu_time: 7.5686 ms
```