@Xreki 2020-11-06
你的profile数据里GatherGPUKernel的时间还是比1.8.5里面时间长一些，确认下是波动、还是别的参数的影响。