@wangchaochaohu 2020-11-06
> 你的profile数据里GatherGPUKernel的时间还是比1.8.5里面时间长一些，确认下是波动、还是别的参数的影响。

测试了三测 分别是1.54 1.53 1.50 应该是数据波动 吧
![image](https://user-images.githubusercontent.com/13411999/98337173-2b858380-2043-11eb-9478-b0096ad92bea.png)
![image](https://user-images.githubusercontent.com/13411999/98337298-5c65b880-2043-11eb-8d1d-d92417d59ef8.png)
![image](https://user-images.githubusercontent.com/13411999/98337319-65ef2080-2043-11eb-8b38-1769bb3df6ef.png)